+++ src/google/bigquery.js
@@ -1,4 +1,0 @@
-const { BigQuery } = require('@google-cloud/bigquery');
-const config = require('../config');
-
-module.exports = new BigQuery(config.gcloud.bigquery);