+++ src/services/elasticsearch.js
@@ -1,0 +1,6 @@
+const config = require('../config');
+const { Client } = require('elasticsearch');
+
+module.exports = new Client({
+  host: config.elasticsearch.server,
+})