+++ src/services/google/datastore.js
@@ -1,4 +1,4 @@
 const Datastore = require('@google-cloud/datastore');
-const config = require('../config');
+const config = require('../../config');
 
 module.exports = new Datastore(config.gcloud.datastore);